backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
Pillow==10.1.0